Correct Viewport Projection + Grid Shader
While testing 3D Rendering, I noticed that the output was flipped on the Y axis. This was due to the viewport being mapped to the OpenGL range. I modified the projected y coordinate to correct this.

Implemented an infinite grid shader following a recipe, and added WASD(QE) input to move the camera in the dev application.